+++ src/main/kotlin/com/mituuz/fuzzier/search/initialview/DefaultInitialListModelProvider.kt
@@ -45,7 +45,7 @@ class DefaultInitialListModelProvider(
             }
 
             FuzzierGlobalSettingsService.RecentFilesMode.RECENTLY_SEARCHED_FILES -> {
-                getRecentlySearchedFiles()
+                getRecentlySearchedFiles(projectState)
             }
 
             else -> {
@@ -82,16 +82,8 @@ class DefaultInitialListModelProvider(
         return listModel
     }
 
-    fun getRecentlySearchedFiles(): DefaultListModel<FuzzyContainer> {
-        val result = DefaultListModel<FuzzyContainer>()
-        projectState.getRecentlySearchedFilesAsFuzzyMatchContainer()
-            .elements()
-            .toList()
-            .filterNotNull()
-            .reversed()
-            .let {
-                result.addAll(it)
-            }
-        return result
-    }
+    fun getRecentlySearchedFiles(state: FuzzierSettingsService.State): DefaultListModel<FuzzyContainer> =
+        DefaultListModel<FuzzyContainer>().apply {
+            state.recentlySearchedFiles?.asReversed()?.forEach { addElement(it.toFuzzyMatchContainer()) }
+        }
 }