+++ src/main/kotlin/com/mituuz/fuzzier/search/initialview/RecentlySearchedFilesUtil.kt
@@ -24,36 +24,75 @@
 
 package com.mituuz.fuzzier.search.initialview
 
+import com.mituuz.fuzzier.entities.FileAccessData
 import com.mituuz.fuzzier.entities.FuzzyContainer
 import com.mituuz.fuzzier.entities.FuzzyMatchContainer
-import com.mituuz.fuzzier.settings.FuzzierGlobalSettingsService
+import com.mituuz.fuzzier.entities.FuzzyMatchContainer.SerializedMatchContainer.Companion.fromFuzzyMatchContainer
 import com.mituuz.fuzzier.settings.FuzzierSettingsService
-import javax.swing.DefaultListModel
 
+/**
+ * Adds a file to the list of recently searched files while ensuring that the list does not exceed
+ * the specified limit, maintains uniqueness, and updates the file metadata cache.
+ *
+ * @param incomingContainer The container representing the file to be added to the recently
+ * searched files list.
+ * @param projectState The current state of the project, which holds the recently searched
+ * files and related metadata.
+ * @param fileListLimit The maximum number of files that can be maintained in the list of
+ * recently searched files.
+ * @param fileMetadataCacheSize The maximum size allowed for the file metadata cache.
+ */
 fun addFileToRecentlySearchedFiles(
-    fuzzyContainer: FuzzyContainer,
+    incomingContainer: FuzzyContainer,
     projectState: FuzzierSettingsService.State,
-    globalState: FuzzierGlobalSettingsService.State
+    fileListLimit: Int,
+    fileMetadataCacheSize: Int,
 ) {
-    val listModel: DefaultListModel<FuzzyMatchContainer> =
-        projectState.getRecentlySearchedFilesAsFuzzyMatchContainer()
+    val recentFiles: MutableList<FuzzyMatchContainer> =
+        projectState.recentlySearchedFiles?.mapNotNull { it.toFuzzyMatchContainer() }?.toMutableList()
+            ?: mutableListOf()
 
     var i = 0
-    while (i < listModel.size) {
-        if (listModel[i].filePath == fuzzyContainer.filePath) {
-            listModel.remove(i)
+    while (i < recentFiles.size) {
+        if (recentFiles[i].filePath == incomingContainer.filePath) {
+            recentFiles.removeAt(i)
         } else {
             i++
         }
     }
 
-    while (listModel.size > globalState.fileListLimit - 1) {
-        listModel.remove(listModel.size - 1)
+    while (recentFiles.size > fileListLimit - 1) {
+        recentFiles.removeAt(recentFiles.size - 1)
     }
 
-    if (fuzzyContainer is FuzzyMatchContainer) {
-        listModel.addElement(fuzzyContainer)
-        projectState.recentlySearchedFiles =
-            FuzzyMatchContainer.SerializedMatchContainer.fromListModel(listModel)
+    if (incomingContainer is FuzzyMatchContainer) {
+        recentFiles.add(incomingContainer)
+        projectState.recentlySearchedFiles = recentFiles.map { fromFuzzyMatchContainer(it) }
     }
+
+    projectState.recentFiles = addFileToLRUCache(
+        incomingContainer, projectState.recentFiles, fileMetadataCacheSize
+    )
+}
+
+fun addFileToLRUCache(
+    incomingContainer: FuzzyContainer, recentFiles: MutableList<FileAccessData>, maxSize: Int
+): MutableList<FileAccessData> {
+    val lower = incomingContainer.filePath.lowercase()
+    val existingIndex = recentFiles.indexOfFirst { it.filePath.lowercase() == lower }
+
+    val existingEntry = if (existingIndex != -1) recentFiles.removeAt(existingIndex) else null
+
+    val newEntry = FileAccessData(
+        filePath = lower,
+        accessCount = (existingEntry?.accessCount ?: 0) + 1
+    )
+
+    recentFiles.add(0, newEntry)
+
+    while (recentFiles.size > maxSize) {
+        recentFiles.removeLast()
+    }
+
+    return recentFiles
 }